+++ NEWS.md
@@ -3,4 +3,5 @@
 ## General
 
 * Replace deprecated `dplyr::add_rownames()` with `tibble::rownames_to_column()`.
-* Improved title labelling for `type = "pred"`.
+* Improved title labelling for `type = "pred"` in `sjp.lm`, `sjp.glm`, `sjp.lmer` and `sjp.glmer`.
+* Improved title and facet title labelling for `type = "eff"` in `sjp.lm`, `sjp.glm`, `sjp.lmer` and `sjp.glmer`.